Add option to do update instead of full build, in which case we just try to pull updates from git/svn and rerun the build process.

+++ install-deps.sh
@@ -1,5 +1,21 @@
 #!/bin/bash
 
+opt_update="false"
+
+# parameters
+until [ -z "$1" ]
+do
+  if [ "$1" == "update" ]; then
+    opt_update="true"
+    echo "Performing update"
+  else
+    echo "Unknown option: $1"
+  fi
+  shift
+done
+
+
+# make sure we have the basic directory layout
 if [ ! -e dependencies ]; then
     mkdir dependencies
 fi
@@ -9,18 +25,24 @@ deps_dir=`pwd`
 
 
 # raknet
-if [ -e raknet ]; then
-  rm -rf raknet
-fi
-if [ -e installed-raknet ]; then
-  rm -rf installed-raknet
+if [ ${opt_update} != "true" ]; then
+  if [ -e raknet ]; then
+    rm -rf raknet
+  fi
+  if [ -e installed-raknet ]; then
+    rm -rf installed-raknet
+  fi
+  mkdir raknet
 fi
 
-mkdir raknet
 cd raknet
-wget http://www.jenkinssoftware.com/raknet/downloads/RakNet-3.51.zip
-unzip RakNet-3.51.zip
-patch -p1 < ../raknet_gcc_4_3.patch
+
+if [ ${opt_update} != "true" ]; then
+  wget http://www.jenkinssoftware.com/raknet/downloads/RakNet-3.51.zip
+  unzip RakNet-3.51.zip
+  patch -p1 < ../raknet_gcc_4_3.patch
+fi
+
 sh bootstrap
 ./configure --prefix=${deps_dir}/installed-raknet
 make
@@ -29,18 +51,23 @@ cd ..
 
 
 # sst
-if [ -e sst ]; then
-  rm -rf sst
-fi
-if [ -e installed-sst ]; then
-  rm -rf installed-sst
+if [ ${opt_update} != "true" ]; then
+  if [ -e sst ]; then
+    rm -rf sst
+  fi
+  if [ -e installed-sst ]; then
+    rm -rf installed-sst
+  fi
+  #svn co svn://svn.pdos.csail.mit.edu/uia/trunk/uia/sst
+  git clone git@ahoy:sst.git
+  cd sst
+  git branch stanford origin/stanford
+  git checkout stanford
+else
+  cd sst
+  git pull origin stanford:stanford
 fi
 
-#svn co svn://svn.pdos.csail.mit.edu/uia/trunk/uia/sst
-git clone git@ahoy:sst.git
-cd sst
-git branch stanford origin/stanford
-git checkout stanford
 misc/setup
 ./configure --prefix=${deps_dir}/installed-sst
 make
@@ -49,19 +76,24 @@ cd ..
 
 
 # sirikata
-if [ -e sirikata ]; then
-  rm -rf sirikata
+if [ ${opt_update} != "true" ]; then
+  if [ -e sirikata ]; then
+    rm -rf sirikata
+  fi
+  if [ -e installed-sirikata ]; then
+    rm -rf installed-sirikata
+  fi
+  git clone git://github.com/sirikata/sirikata.git sirikata
+  cd sirikata
+else
+  cd sirikata
+  git pull origin
 fi
-if [ -e installed-sirikata ]; then
-  rm -rf installed-sirikata
-fi
-
-git clone git://github.com/sirikata/sirikata.git sirikata
-cd sirikata
 make depends
 git submodule init
 git submodule update
 cd build/cmake
 cmake -DCMAKE_INSTALL_PREFIX=${deps_dir}/installed-sirikata .
 make -j2
+make install
 cd ../../..
